requirements.txt
django-phone-login==1.1.3
djangorestframework==3.14.0